requirements.txt
streamlit
opaiui
python-dotenv